+++ src/kernel/allocator/allocator.rs
@@ -28,16 +28,17 @@ use crate::{
 pub const HEAP_SIZE: usize = 1024 * 1024; // 1 MB heap size
 
 // defining the Allocator (which implements the 'GlobalAlloc' trait)
+#[cfg(feature = "global-alloc")] // Defaultfeature für Kernel deaktiviert -> allocator Dopplung
 #[global_allocator]
-//static ALLOCATOR: Locked<BumpAllocator> = Locked::new(BumpAllocator::new());
 static ALLOCATOR: Locked<LinkedListAllocator> = Locked::new(LinkedListAllocator::new());
 
 /**
  Description: Initialization of the allocator. Must be called early in 'startup'.
 */
+#[cfg(feature = "global-alloc")] // Defaultfeature für Kernel deaktiviert -> allocator Dopplung
 pub fn init(pid: usize, heap_size: usize) {
     // Erst speicher anfordern
-    let heap_start: usize = usr_mmap_heap_space(pid, heap_size as u64) as usize;
+    let heap_start: usize = usr_mmap_heap_space(pid, heap_size) as usize;
 
     // Fehlerprüfung
     if heap_start == 0 {
@@ -54,6 +55,7 @@ pub fn init(pid: usize, heap_size: usize) {
  Description: Allocates memory from the heap.
              Compiler generates code calling this function.
 */
+#[cfg(feature = "global-alloc")] // Defaultfeature für Kernel deaktiviert -> allocator Dopplung
 pub fn alloc(layout: Layout) -> *mut u8 {
     unsafe { ALLOCATOR.lock().alloc(layout) }
 }
@@ -62,6 +64,7 @@ pub fn alloc(layout: Layout) -> *mut u8 {
  Description: Deallocates memory from the heap.
              Compiler generates code calling this function.
 */
+#[cfg(feature = "global-alloc")] // Defaultfeature für Kernel deaktiviert -> allocator Dopplung
 pub fn dealloc(ptr: *mut u8, layout: Layout) {
     unsafe { ALLOCATOR.lock().dealloc(ptr, layout) }
 }
